Fix Float vs CGFloat type mismatches to match Apple API
- CAEmitterCell: 7 properties (velocity, velocityRange, scale,
  scaleRange, scaleSpeed, spin, spinRange) Float → CGFloat
- CAGradientLayer: locations [Float]? → [CGFloat]?
- CAKeyframeAnimation: keyTimes, tensionValues, continuityValues,
  biasValues [Float]? → [CGFloat]?
- CALayer: internal keyTimes/locations calculations use CGFloat
- CLAUDE.md: add type compatibility and Float/CGFloat usage rules

Closes #1

+++ Sources/OpenCoreAnimation/CAGradientLayer.swift
@@ -43,10 +43,10 @@ open class CAGradientLayer: CALayer {
         }
     }
 
-    internal var _locations: [Float]?
+    internal var _locations: [CGFloat]?
     /// An optional array of numbers defining the location of each gradient stop. Animatable.
     /// Values should be in the range [0, 1] and monotonically increasing.
-    open var locations: [Float]? {
+    open var locations: [CGFloat]? {
         get { return _locations }
         set {
             let oldValue = _locations
